+++ itemsell.php
@@ -64,7 +64,7 @@ if (!empty($_POST['qty']) && !empty($_GET['ID']))
         }
         else
         {
-            $price = $r['itmsellprice'] * $_POST['qty'];
+            $price = (int)($r['itmsellprice'] * $_POST['qty']);
             item_remove($userid, $r['itmid'], $_POST['qty']);
             $db->query(
                     "UPDATE `users`
